[INTG-1521] Add integration tests for SQL databases
This PR adds integration tests for SQL databases. You might notice that these look like they are testing CSV exporting... Well that is because they are! These test leverge the fact that our CSV exporter wraps our SQLExporter and uses a file based SQLite DB to perform incremental refreshes.

Instead of using that SQLite DB, we just use a real one. Simple as that, we have tested our ability to export to a SQL db, and comparing the output is as easy as comparing 2 CSV files.
